# Response refactor, increase requests-per-second, body type handling