### Description
This PR refactors response generation, speeding up operations with various body types and sizes.  It also adds code to better handle the various body types returned by Rails and/or Rack.  Much of this has been worked on by @guilleiguaran, so thanks for his work on it (see #2892).  Essentially, body types fall into five categories:

1. An Array - Puma will determine the content-length if not provided.
2. An Enum - this would be chunked by Puma or sent as is if chunked encoding has already been applied.
3. A File object - dependent on its size, it will either be sent with the headers, or sent after the headers using `IO.copy_stream`.
4. An object that responds to `to_path` - Puma will attempt to open the file, and if successful, will send as above.
5. An object that responds to `call`.

This PR has been around for a while, and I'm sure I've run millions of requests thru it.  Two items:

1. Duplicating client errors is difficult.  I believe I've got the error handling correct, but we may see some issues with it.  Not.sure.
2. Previously, Puma didn't ever call `to_path` on a response body.  This PR has been updated to do so when the body responds to it, but isn't a File object.  It opens the file (based on the name in `to_path`), and treats it as a File/IO object.  This may cause issues with the [Rack::Sendfile](https://msp-greg.github.io/rack/Rack/Sendfile.html) middleware.

Below summarizes performance using the code in benchmarks/local, all using:
benchmarks/local/response_time_wrk.sh -w2 -t5:5 -s tcp6 -Y
wrk -t8 -c16 -d10s
Server cluster mode -w2 -t5:5, bind: tcp6
ruby 3.2.0dev (2022-09-12T13:13:32Z master 2aa8edaec7) +YJIT [x86_64-linux]
Both include 'test/rackup/ci_*.ru files - cache response bodies' (#2952)
Windows WSL2/Ubuntu 22.04

```

Body    ────────── req/sec ──────────   ─────── req 50% times ─────── Master
 KB     array   chunk  string      io   array   chunk  string      io
1       13780    9719   13581   10964    0.65    1.22    0.68    0.90
10       5920    2769   13203   10831    1.78    4.04    0.80    0.95
100      1063     404   11002    9153   10.54   26.85    0.98    1.21
256       448     159   11264    9039   24.39   65.65    0.92    1.23
512       234      85   10757    8848   45.18  121.76    0.96    1.22
1024      129      43    8822    8543   79.86  237.58    0.77    1.21
2048       68      22    6342    8405  152.13  456.80    1.54    1.21
─────────────────────────────────────────────────────────────────────

Body    ────────── req/sec ──────────   ─────── req 50% times ─────── PR #2896
 KB     array   chunk  string      io   array   chunk  string      io
1       15889   15885   15594   10898   0.200   0.197   0.195   0.960
10      15419   15204   15605   11073   0.297   0.285   0.225   0.728
100     10596    9660   11324    9396   1.020   1.120   0.960   1.180
256      9198    7501   11960    9286   1.120   1.360   0.860   1.190
512      7112    5512   11012    9017   1.430   1.870   0.930   1.190
1024     4355    3315    9302    8762   1.740   2.760   0.709   1.170
2048     2801    1938    6595    8562   3.460   5.240   1.470   1.170
─────────────────────────────────────────────────────────────────────
```

There are significant speed increases for almost all body types/sizes, with the exception of 'single string' and 'io/file' bodies, which show smaller increases.

The commits:
* [io_buffer.rb - change to using StringIO](https://github.com/puma/puma/commit/8ca473f7b0c0da4d1a3bda06be0d9500e5dbe5f1) - allows IOBuffer to be used both as a string and as an IO.  Previously it was subclassed from `String`.

* [request.rb - refactor](https://github.com/puma/puma/commit/75837c456295e69713e2247b3e46004bdb0391a3) - Split `fast_write` into `fast_write_response` and `fast_write_str`.  Refactor.  Compute array bodies' `Content-Length` only when it's not provided by app.  See #2892 by @guilleiguaran.  Only an Array responds to `#to_ary`, so it is used for the body type check.

* [client.rb - small refactor](https://github.com/puma/puma/commit/c2cd372aa858f4d976709b7d917b26b4ad35d898)

All three of the below are small refactors

* [test_persistent.rb - chunked tests must be enum, not array](https://github.com/puma/puma/commit/0962c28596a354cf7360e65440f7b2f57a67a36b)
* [test_rack_server.rb - Add Rack::BodyProxy tests](https://github.com/puma/puma/commit/c1e6028d971093459d798c99e431a8d1c91c31dc) - from PR #2892.
* [test_puma_server.rb - always start one thread](https://github.com/puma/puma/commit/3dd99b064210e108d226b23fe4762bdb1fecb8af)

* [request.rb - rename variables](https://github.com/puma/puma/pull/2896/commits/1d58951da1d4) - Rename `lines` to `io_buffer`, which is what it is.  Rename `io` to `socket`.

Most of the remaining commits involve code to handle various body types, along with tests.

Please feel free to test these changes.

Closes #2892, closes #2703, closes #2457

### Your checklist for this pull request
- [x] I have reviewed the [guidelines for contributing](../blob/master/CONTRIBUTING.md) to this repository.
- [ ] I have added (or updated) appropriate tests if this PR fixes a bug or adds a feature.
- [ ] My pull request is 100 lines added/removed or less so that it can be easily reviewed.
- [ ] If this PR doesn't need tests (docs change), I added `[ci skip]` to the title of the PR.
- [x] If this closes any issues, I have added "Closes `#issue`" to the PR description or my commit messages.
- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed, including Rubocop.