@JuanitoFatas 2022-09-13
Late to the party, but I tried this on [Ruby 3.2.0 Preview 2](https://www.ruby-lang.org/en/news/2022/09/09/ruby-3-2-0-preview2-released/). Following are the results.

#### My laptop information

```
ProductName:	macOS
ProductVersion:	12.4
BuildVersion:	21F79
ruby 3.2.0preview2 (2022-09-09 master 35cfc9a3bb) [arm64-darwin21] without YJIT
```


#### master - 0e832e71ef

```
$ benchmarks/local/response_time_wrk.sh -w2 -t5:5 -s ssl6 -Y

Body    ────────── req/sec ──────────   ─────── req 50% times ───────
 KB     array   chunk  string      io   array   chunk  string      io
1       20293   17969   22868   18407    0.36    0.60    0.53    0.57
10      11181    4149   19329   16614    0.74    2.19    0.42    0.65
100      1036     236    8728    7782    8.91   69.89    1.03    1.37
256       905     272    4574    4199    7.71   21.60    1.89    2.37
512       431     149    2729    2326   15.20   41.19    3.55    4.01
1024      172      30    1428    1261   35.52  103.26    6.54    7.58
2048       47      16     759     632  155.84  646.02   12.67   15.15
─────────────────────────────────────────────────────────────────────
wrk -t8 -c16 -d10s
benchmarks/local/response_time_wrk.sh -w2 -t5:5 -s ssl6 -Y
Server cluster mode -w2 -t5:5, bind: ssl6
Puma repo branch master
ruby 3.2.0preview2 (2022-09-09 master 35cfc9a3bb) [arm64-darwin21]
```

<details>
<summary>View Full Result</summary>

```
$ benchmarks/local/response_time_wrk.sh -w2 -t5:5 -s ssl6 -Y

bundle exec bin/puma -q -b ssl://[::1]:40001?cert=examples/puma/cert_puma.pem&key=examples/puma/puma_keypair.pem&verify_mode=none -S tmp/bench_test_puma.state -w2 -t5:5 --control-url=tcp://127.0.0.1:40010 --control-token=test test/rackup/ci_select.ru

usage: sleep seconds
ruby: warning: Ruby was built without YJIT support
ruby: warning: Ruby was built without YJIT support
[14328] Puma starting in cluster mode...
[14328] * Puma version: 5.6.5 (ruby 3.2.0-p-1) ("Birdie's Version")
[14328] *  Min threads: 5
[14328] *  Max threads: 5
[14328] *  Environment: development
[14328] *   Master PID: 14328
[14328] *      Workers: 2
[14328] *     Restarts: (✔) hot (✔) phased
[14328] * Listening on ssl://[::1]:40001?cert=examples/puma/cert_puma.pem&key=examples/puma/puma_keypair.pem&verify_mode=none
[14328] Use Ctrl-C to stop
[14328] * Starting control server on http://127.0.0.1:40010
[14328] - Worker 1 (PID: 14333) booted in 0.0s, phase: 0
[14328] - Worker 0 (PID: 14332) booted in 0.01s, phase: 0

warm-up
wrk -t2 -c4 -d1s --latency -H 'Body-Conf: a10'   https://[::1]:40001 |  13918 requests in  1.10s, 159.46MB read
wrk -t2 -c4 -d1s --latency -H 'Body-Conf: c10'   https://[::1]:40001 |   6163 requests in  1.10s,  71.10MB read
wrk -t2 -c4 -d1s --latency -H 'Body-Conf: s10'   https://[::1]:40001 |  22974 requests in  1.10s, 263.20MB read
wrk -t2 -c4 -d1s --latency -H 'Body-Conf: i10'   https://[::1]:40001 |  18741 requests in  1.10s, 214.71MB read

wrk -t8 -c16 -d10s --latency -H 'Body-Conf: a1'    https://[::1]:40001 | 204979 requests in 10.10s, 546.71MB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: c1'    https://[::1]:40001 | 181543 requests in 10.10s, 487.25MB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: s1'    https://[::1]:40001 | 230961 requests in 10.10s, 615.87MB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: i1'    https://[::1]:40001 | 185924 requests in 10.10s, 495.81MB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: a10'   https://[::1]:40001 | 112923 requests in 10.10s,   1.26GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: c10'   https://[::1]:40001 |  41905 requests in 10.10s, 483.31MB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: s10'   https://[::1]:40001 | 195247 requests in 10.10s,   2.18GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: i10'   https://[::1]:40001 | 167796 requests in 10.10s,   1.88GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: a100'  https://[::1]:40001 |  10461 requests in 10.10s,   1.02GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: c100'  https://[::1]:40001 |   2371 requests in 10.03s, 237.69MB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: s100'  https://[::1]:40001 |  88165 requests in 10.10s,   8.55GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: i100'  https://[::1]:40001 |  78638 requests in 10.10s,   7.63GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: a256'  https://[::1]:40001 |   9139 requests in 10.10s,   2.25GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: c256'  https://[::1]:40001 |   2743 requests in 10.10s, 696.48MB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: s256'  https://[::1]:40001 |  46207 requests in 10.10s,  11.36GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: i256'  https://[::1]:40001 |  42425 requests in 10.10s,  10.43GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: a512'  https://[::1]:40001 |   4351 requests in 10.10s,   2.13GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: c512'  https://[::1]:40001 |   1509 requests in 10.10s, 764.44MB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: s512'  https://[::1]:40001 |  27568 requests in 10.10s,  13.51GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: i512'  https://[::1]:40001 |  23491 requests in 10.10s,  11.51GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: a1024' https://[::1]:40001 |   1733 requests in 10.10s,   1.70GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: c1024' https://[::1]:40001 |    298 requests in 10.10s, 304.86MB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: s1024' https://[::1]:40001 |  14428 requests in 10.10s,  14.12GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: i1024' https://[::1]:40001 |  12739 requests in 10.10s,  12.47GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: a2048' https://[::1]:40001 |    473 requests in 10.04s,   0.94GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: c2048' https://[::1]:40001 |    158 requests in 10.10s, 325.42MB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: s2048' https://[::1]:40001 |   7668 requests in 10.10s,  15.00GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: i2048' https://[::1]:40001 |   6388 requests in 10.10s,  12.50GB read

Type   req/sec    50%     75%     90%     99%    100%  Resp Size  Errors
───────────────────────────────────────────────────────────────────────────────────    1kB
array   20293     0.4     0.5     0.8     3.4    11.0      1023
chunk   17969     0.6     1.1     2.8     5.4    11.6      1040
string  22868     0.5     1.0     2.4     4.3    13.2      1022
io      18407     0.6     0.9     3.2     6.0    15.2      1022
───────────────────────────────────────────────────────────────────────────────────   10kB
array   11181     0.7     1.0     1.3     4.4    15.8     10239
chunk    4149     2.2     7.1    19.5    56.0    96.4     10319
string  19329     0.4     0.6     1.2     4.3    12.1     10238
io      16614     0.7     1.1     3.2     5.9    15.6     10239
───────────────────────────────────────────────────────────────────────────────────  100kB
array    1036     8.9    23.5    72.0   214.2   313.9    102444
chunk     236    69.9   100.9   275.5   496.6   638.2    103344
string   8728     1.0     1.4     2.5     9.4    21.4    102400
io       7782     1.4     2.2     8.0    14.6    28.2    102400
───────────────────────────────────────────────────────────────────────────────────  256kB
array     905     7.7    15.5    25.4    51.5   286.1    262277
chunk     272    21.6    64.8   103.7   337.2  1050.0    264470
string   4574     1.9     2.4     2.9     7.2    28.5    262109
io       4199     2.4     3.1     9.0    20.6    32.4    262148
───────────────────────────────────────────────────────────────────────────────────  512kB
array     431    15.2    37.5    59.6   155.7   473.8    524888
chunk     149    41.2   110.0   174.3   365.3  1010.0    529420
string   2729     3.6     4.3     7.1    36.7    47.4    524522
io       2326     4.0     4.7     5.4    12.3    54.6    524349
─────────────────────────────────────────────────────────────────────────────────── 1024kB
array     172    35.5   100.7   140.7   869.0  1710.0   1051747
chunk      30   103.3   708.9   799.9   913.7  1470.0   1070940
string   1428     6.5     7.9     9.6    26.2    67.5   1049204
io       1261     7.6     8.5     9.5    18.6    73.1   1048938
─────────────────────────────────────────────────────────────────────────────────── 2048kB
array      47   155.8   425.9   523.4  1080.0  1580.0   2121956
chunk      16   646.0   806.9   888.2  1110.0  1160.0   2157877
string    759    12.7    14.6    16.6    31.6   140.6   2098010
io        632    15.2    17.0    20.4    36.0   173.2   2099701

Body    ────────── req/sec ──────────   ─────── req 50% times ───────
 KB     array   chunk  string      io   array   chunk  string      io
1       20293   17969   22868   18407    0.36    0.60    0.53    0.57
10      11181    4149   19329   16614    0.74    2.19    0.42    0.65
100      1036     236    8728    7782    8.91   69.89    1.03    1.37
256       905     272    4574    4199    7.71   21.60    1.89    2.37
512       431     149    2729    2326   15.20   41.19    3.55    4.01
1024      172      30    1428    1261   35.52  103.26    6.54    7.58
2048       47      16     759     632  155.84  646.02   12.67   15.15
─────────────────────────────────────────────────────────────────────
wrk -t8 -c16 -d10s
benchmarks/local/response_time_wrk.sh -w2 -t5:5 -s ssl6 -Y
Server cluster mode -w2 -t5:5, bind: ssl6
Puma repo branch master
ruby 3.2.0preview2 (2022-09-09 master 35cfc9a3bb) [arm64-darwin21]

[14328] - Gracefully shutting down workers...
[14328] === puma shutdown: 2022-09-13 20:58:46 +0900 ===
[14328] - Goodbye!

 5:14 Total Time
```

</details>

#### This PR - c32f8ff9b

```
$ benchmarks/local/response_time_wrk.sh -w2 -t5:5 -s ssl6 -Y

Body    ────────── req/sec ──────────   ─────── req 50% times ───────
 KB     array   chunk  string      io   array   chunk  string      io
1       23522   24581   25445   21879   0.405   0.424   0.455   0.537
10      22458   21593   22811   19156   0.472   0.430   0.487   0.568
100      7772    7431    8768    6853   1.100   1.330   1.230   1.210
256      4089    3646    4731    4037   2.150   2.410   2.050   2.210
512      2303    2021    2474    2244   3.960   4.560   3.730   4.100
1024     1170    1075    1442    1262   7.930   8.840   6.510   7.570
2048      565     538     755     656  16.640  17.940  12.770  14.790
─────────────────────────────────────────────────────────────────────
wrk -t8 -c16 -d10s
benchmarks/local/response_time_wrk.sh -w2 -t5:5 -s ssl6 -Y
Server cluster mode -w2 -t5:5, bind: ssl6
Puma repo branch 00-response-refactor
ruby 3.2.0preview2 (2022-09-09 master 35cfc9a3bb) [arm64-darwin21]
```

<details>
<summary>View Full Result</summary>

```
$ benchmarks/local/response_time_wrk.sh -w2 -t5:5 -s ssl6 -Y

bundle exec bin/puma -q -b ssl://[::1]:40001?cert=examples/puma/cert_puma.pem&key=examples/puma/puma_keypair.pem&verify_mode=none -S tmp/bench_test_puma.state -w2 -t5:5 --control-url=tcp://127.0.0.1:40010 --control-token=test test/rackup/ci_select.ru

usage: sleep seconds
ruby: warning: Ruby was built without YJIT support
ruby: warning: Ruby was built without YJIT support
[13614] Puma starting in cluster mode...
[13614] * Puma version: 5.6.5 (ruby 3.2.0-p-1) ("Birdie's Version")
[13614] *  Min threads: 5
[13614] *  Max threads: 5
[13614] *  Environment: development
[13614] *   Master PID: 13614
[13614] *      Workers: 2
[13614] *     Restarts: (✔) hot (✔) phased
[13614] * Listening on ssl://[::1]:40001?cert=examples/puma/cert_puma.pem&key=examples/puma/puma_keypair.pem&verify_mode=none
[13614] Use Ctrl-C to stop
[13614] * Starting control server on http://127.0.0.1:40010
[13614] - Worker 0 (PID: 13628) booted in 0.01s, phase: 0
[13614] - Worker 1 (PID: 13629) booted in 0.01s, phase: 0

warm-up
wrk -t2 -c4 -d1s --latency -H 'Body-Conf: a10'   https://[::1]:40001 |  26562 requests in  1.10s, 304.31MB read
wrk -t2 -c4 -d1s --latency -H 'Body-Conf: c10'   https://[::1]:40001 |  26710 requests in  1.10s, 308.04MB read
wrk -t2 -c4 -d1s --latency -H 'Body-Conf: s10'   https://[::1]:40001 |  26873 requests in  1.10s, 307.87MB read
wrk -t2 -c4 -d1s --latency -H 'Body-Conf: i10'   https://[::1]:40001 |  21684 requests in  1.10s, 248.42MB read

wrk -t8 -c16 -d10s --latency -H 'Body-Conf: a1'    https://[::1]:40001 | 237588 requests in 10.10s, 633.59MB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: c1'    https://[::1]:40001 | 248277 requests in 10.10s, 666.33MB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: s1'    https://[::1]:40001 | 257024 requests in 10.10s, 685.37MB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: i1'    https://[::1]:40001 | 220993 requests in 10.10s, 589.28MB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: a10'   https://[::1]:40001 | 226841 requests in 10.10s,   2.54GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: c10'   https://[::1]:40001 | 218111 requests in 10.10s,   2.46GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: s10'   https://[::1]:40001 | 230415 requests in 10.10s,   2.58GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: i10'   https://[::1]:40001 | 193550 requests in 10.10s,   2.17GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: a100'  https://[::1]:40001 |  78517 requests in 10.10s,   7.62GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: c100'  https://[::1]:40001 |  75054 requests in 10.10s,   7.33GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: s100'  https://[::1]:40001 |  88591 requests in 10.10s,   8.60GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: i100'  https://[::1]:40001 |  69228 requests in 10.10s,   6.72GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: a256'  https://[::1]:40001 |  41302 requests in 10.10s,  10.15GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: c256'  https://[::1]:40001 |  36824 requests in 10.10s,   9.11GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: s256'  https://[::1]:40001 |  47792 requests in 10.10s,  11.75GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: i256'  https://[::1]:40001 |  40768 requests in 10.10s,  10.02GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: a512'  https://[::1]:40001 |  23257 requests in 10.10s,  11.40GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: c512'  https://[::1]:40001 |  20412 requests in 10.10s,  10.07GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: s512'  https://[::1]:40001 |  24992 requests in 10.10s,  12.25GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: i512'  https://[::1]:40001 |  22664 requests in 10.10s,  11.11GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: a1024' https://[::1]:40001 |  11815 requests in 10.10s,  11.56GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: c1024' https://[::1]:40001 |  10864 requests in 10.10s,  10.70GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: s1024' https://[::1]:40001 |  14565 requests in 10.10s,  14.25GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: i1024' https://[::1]:40001 |  12749 requests in 10.10s,  12.47GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: a2048' https://[::1]:40001 |   5656 requests in 10.01s,  11.06GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: c2048' https://[::1]:40001 |   5433 requests in 10.10s,  10.70GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: s2048' https://[::1]:40001 |   7628 requests in 10.10s,  14.92GB read
wrk -t8 -c16 -d10s --latency -H 'Body-Conf: i2048' https://[::1]:40001 |   6625 requests in 10.10s,  12.96GB read

Type   req/sec    50%     75%     90%     99%    100%  Resp Size
─────────────────────────────────────────────────────────────────    1kB
array   23522    0.40    0.63    1.66    3.50   10.67      1022
chunk   24581    0.42    0.66    2.27    4.34   11.19      1040
string  25445    0.46    0.76    2.06    3.75   13.53      1022
io      21879    0.54    0.95    3.15    5.15   13.91      1022
─────────────────────────────────────────────────────────────────   10kB
array   22458    0.47    0.74    2.02    3.77   11.16     10238
chunk   21593    0.43    0.65    1.69    4.07   16.12     10318
string  22811    0.49    0.79    2.20    4.11   11.73     10238
io      19156    0.57    0.96    2.70    5.07   16.60     10238
─────────────────────────────────────────────────────────────────  100kB
array    7772    1.10    1.50    2.18   11.60   28.77    102403
chunk    7431    1.33    1.89    5.60   12.58   22.79    103110
string   8768    1.23    2.04    5.79   10.36   18.90    102445
io       6853    1.21    1.58    2.02    6.84   26.84    102406
─────────────────────────────────────────────────────────────────  256kB
array    4089    2.15    2.67    3.22    6.04   32.39    262218
chunk    3646    2.41    3.00    3.61    7.66   30.20    263806
string   4731    2.05    2.65    6.64   21.66   34.47    262144
io       4037    2.21    2.69    3.22    7.24   30.38    262079
─────────────────────────────────────────────────────────────────  512kB
array    2303    3.96    4.75    5.62   11.08   57.11    524597
chunk    2021    4.56    5.54    6.55   11.35   57.56    528084
string   2474    3.73    4.57    5.40    9.28   42.29    524419
io       2244    4.10    4.88    5.85   10.87   61.56    524574
───────────────────────────────────────────────────────────────── 1024kB
array    1170    7.93    9.37   11.32   21.85   95.88   1048425
chunk    1075    8.84   10.37   12.14   22.93   95.74   1055826
string   1442    6.51    7.65    8.91   21.50   60.49   1048690
io       1262    7.57    8.53    9.59   24.34   94.99   1048879
───────────────────────────────────────────────────────────────── 2048kB
array     565   16.64   20.12   24.35   38.16  176.99   2098685
chunk     538   17.94   21.14   24.58   35.24  184.88   2113056
string    755   12.77   14.66   16.75   32.02  155.45   2098783
io        656   14.79   16.34   18.30   37.37  188.06   2098245

Body    ────────── req/sec ──────────   ─────── req 50% times ───────
 KB     array   chunk  string      io   array   chunk  string      io
1       23522   24581   25445   21879   0.405   0.424   0.455   0.537
10      22458   21593   22811   19156   0.472   0.430   0.487   0.568
100      7772    7431    8768    6853   1.100   1.330   1.230   1.210
256      4089    3646    4731    4037   2.150   2.410   2.050   2.210
512      2303    2021    2474    2244   3.960   4.560   3.730   4.100
1024     1170    1075    1442    1262   7.930   8.840   6.510   7.570
2048      565     538     755     656  16.640  17.940  12.770  14.790
─────────────────────────────────────────────────────────────────────
wrk -t8 -c16 -d10s
benchmarks/local/response_time_wrk.sh -w2 -t5:5 -s ssl6 -Y
Server cluster mode -w2 -t5:5, bind: ssl6
Puma repo branch 00-response-refactor
ruby 3.2.0preview2 (2022-09-09 master 35cfc9a3bb) [arm64-darwin21]

[13614] - Gracefully shutting down workers...
[13614] === puma shutdown: 2022-09-13 20:50:47 +0900 ===
[13614] - Goodbye!

 5:14 Total Time
```

</details>